Adições e Alterações em ProfileController e UserController, Middlewares, resources/views/layouts e route/web

+++ routes/web.php
@@ -19,12 +19,12 @@ foreach(File::allFiles(__DIR__.'/web') as $route_file){
 //Ofertas
 Route::prefix('ofertas')->group(function () {
     Route::get('/index', [OfertaController::class, 'index'])->middleware(['auth'])->name('ofertas.index');
-    Route::get('/create', [OfertaController::class, 'create'])->middleware(['auth'])->name('ofertas.create');
-    Route::post('/store', [OfertaController::class, 'store'])->middleware(['auth'])->name('ofertas.store');
+    Route::get('/create', [OfertaController::class, 'create'])->middleware(['auth','cliente'])->name('ofertas.create');
+    Route::post('/store', [OfertaController::class, 'store'])->middleware(['auth','cliente'])->name('ofertas.store');
     Route::get('/show/{id}',[OfertaController::class,'show'])->middleware(['auth'])->name('ofertas.show');
-    Route::get('/edit/{id}', [OfertaController::class, 'edit'])->middleware(['auth'])->name('ofertas.edit');
-    Route::post('/update/{id}', [OfertaController::class, 'update'])->middleware(['auth'])->name('ofertas.update');
-    Route::delete('/delete/{id}', [OfertaController::class, 'destroy'])->middleware(['auth'])->name('ofertas.delete');
+    Route::get('/edit/{id}', [OfertaController::class, 'edit'])->middleware(['auth','clienteOrAdmin'])->name('ofertas.edit');
+    Route::post('/update/{id}', [OfertaController::class, 'update'])->middleware(['auth','clienteOrAdmin'])->name('ofertas.update');
+    Route::delete('/delete/{id}', [OfertaController::class, 'destroy'])->middleware(['auth','clienteOrAdmin'])->name('ofertas.delete');
 });
 
 require __DIR__.'/auth.php';
